+++ packages/twenty-shared/src/utils/indexMetadata/getUniqueConstraintsFields.ts
@@ -1,11 +1,22 @@
-import { isDefined } from 'twenty-shared/utils';
+import { isDefined } from '@/utils/validation/isDefined';
 
-import { FieldMetadataInterface } from 'src/engine/metadata-modules/field-metadata/interfaces/field-metadata.interface';
-import { ObjectMetadataInterface } from 'src/engine/metadata-modules/field-metadata/interfaces/object-metadata.interface';
-
-export const getUniqueConstraintsFields = (
-  objectMetadata: ObjectMetadataInterface,
-): FieldMetadataInterface[][] => {
+export const getUniqueConstraintsFields = <
+  K extends {
+    id: string;
+    name: string;
+  },
+  T extends {
+    id: string;
+    indexMetadatas: {
+      id: string;
+      isUnique: boolean;
+      indexFieldMetadatas: { fieldMetadataId: string }[];
+    }[];
+    fields: K[];
+  },
+>(
+  objectMetadata: T,
+): K[][] => {
   const uniqueIndexes = objectMetadata.indexMetadatas.filter(
     (index) => index.isUnique,
   );
